Change column of tables cart and cart_items to decimal

+++ db/migrate/20170525023647_create_carts.rb
@@ -2,10 +2,10 @@ class CreateCarts < ActiveRecord::Migration[5.1]
   def change
     create_table :carts do |t|
       t.references :user, foreign_key: true
-      t.integer :igv
-      t.integer :discount
-      t.integer :subtotal
-      t.integer :total
+      t.decimal :igv
+      t.decimal :discount
+      t.decimal :subtotal
+      t.decimal :total
 
       t.timestamps
     end
